Remove dependency on native-keymap
The native-keymap module did not work as expected on windows: the returned
characters did not respect the current keyboard layout. We have discovered
that we only really need an explicit keymap on Mac anyway, in order
to differentiate keystrokes using option as a modifier from keystrokes using
option to insert a different character. On linux and windows, we can
tell the difference between these two situations by querying the `AltGraph`
modifier state of the keyboard event.

Signed-off-by: Nathan Sobo <nathan@github.com>